back/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.68.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==1.10.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.32.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==1.4.23
sqlparse==0.5.3
starlette==0.14.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.15.0
